fix: datetime2 rounding in mssql (#7264)
This upgrades the `mssql` driver to version 7 to resolve #3202. This
issue was created in the bulk insert test adding a datetime2 field
with a time of midnight.

v7 no longer allows numeric parameter names and was throwing an error
inside tedious but you can see that `node-mssql` is expecting a string.
https://github.com/tediousjs/node-mssql/blob/76585f973dd8cad48836fe302f99f67d47ff6129/lib/base/request.js#L101
Another user saw this issue in the comments of issue #3202.
https://github.com/typeorm/typeorm/issues/3202#issuecomment-734926616

Stringifying the numeric parameter names before it enters the driver
resolves this issue.

Now all mssql tests pass locally once you provide additional config in
`ormconfig.json` locally required for a locally running docker instance
of mssql which uses a self signed certificate.
```json
"extra": {
  "trustServerCertificate": true
}
```